add flow control between server and agent
On the agent side, there is a buffered channel between the tunnel and the
outgoing connection. If this side is blocking and the buffer (5) is
filled the tunnel's send on the connection channel blocks and no more packets
from the server side can be accepted anymore. As a result not only the
blocking connection stops, but the complete set of outgoing connections
do not transfer data anymore.

On the server side there is even no such buffer between the tunnel connection
and the connection. Every packet received from the tunnel is synchronously
written to the connection (either a grpc frontened or hijacked http tcp
connection).

In any case one blocking connection can block the complete tunnel. Therefore
a flow control is required. On both sides a buffer is required for the
outgoing connection traffic and the sending side must stop sending packets
into the tunnel, if there is no capacity in the buffer anymore.

This is achieved by introducing a DATA_ACK message and semaphore with the
capacity according to the accepting side's buffer size. This semaphore
limits the packet send operations by consuming a token before the send
operation and releasing a token for every ACK.